+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/entries/ExecutionTests.cs
@@ -17,4 +17,11 @@ public class ExecutionTests : InMemoryExecutionTestsBase
     [Fact(DisplayName = "iteration.js")]
     public Task iteration()
         => ExecutionTestFromFile("iteration");
+
+    [Fact(DisplayName = "resizable-buffer-grow-mid-iteration.js")]
+    public Task resizable_buffer_grow_mid_iteration() => ExecutionTestFromFile("resizable-buffer-grow-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer-shrink-mid-iteration.js")]
+    public Task resizable_buffer_shrink_mid_iteration() => ExecutionTestFromFile("resizable-buffer-shrink-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }