+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/map/ExecutionTests.cs
@@ -227,4 +227,10 @@ public partial class ExecutionTests : InMemoryExecutionTestsBase
     public Task _15_4_4_19_4_15()
         => ExecutionTestFromFile("15.4.4.19-4-15");
 
+    [Fact(DisplayName = "resizable-buffer-grow-mid-iteration.js")]
+    public Task resizable_buffer_grow_mid_iteration() => ExecutionTestFromFile("resizable-buffer-grow-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer-shrink-mid-iteration.js")]
+    public Task resizable_buffer_shrink_mid_iteration() => ExecutionTestFromFile("resizable-buffer-shrink-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }