+++ tests/Jroc.Test262.Tests/built-ins/TypedArray/prototype/values/ExecutionTests.cs
@@ -9,4 +9,11 @@ public class ExecutionTests : DiskExecutionTestsBase
         [Fact(DisplayName = "iter-prototype")]
     public Task iter_prototype()
         => ExecutionTestFromFile("iter-prototype");
+
+    [Fact(DisplayName = "resizable-buffer-grow-mid-iteration.js")]
+    public Task resizable_buffer_grow_mid_iteration() => ExecutionTestFromFile("resizable-buffer-grow-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer-shrink-mid-iteration.js")]
+    public Task resizable_buffer_shrink_mid_iteration() => ExecutionTestFromFile("resizable-buffer-shrink-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }